+++ ChangeLog.md
@@ -13,3 +13,9 @@
 
 ### Changed
 - Improved documentation for all exposed modules.
+
+## 1.1.3  -- 2018-07-30
+
+### Changed
+- Fixed a major bug with persistent homology; high dimensional holes were being detected in low dimensional data sets.
+- Persistent homology now filters out bar codes of the form (i, Just i), as they say nothing about the topology of the underlying complexes.